+++ src/os2datascanner/projects/report/reportapp/urls.py
@@ -1,16 +1,17 @@
 import django.contrib.auth.views as auth_views
-from django.urls import re_path
+from django.urls import re_path, include, path
 from django.http import HttpResponse
-from django.urls import include
 from django.conf.urls.static import static
 from django.conf import settings
-from django.urls import path
 from django.views.i18n import JavaScriptCatalog
+from django.views.generic import RedirectView
+
 from os2datascanner import __version__
 
 from .views.api import JSONAPIView
 from .views.statistics_views import (
-    LeaderStatisticsPageView, LeaderStatisticsCSVView, DPOStatisticsPageView, DPOStatisticsCSVView,
+    LeaderStatisticsPageView, LeaderStatisticsCSVView,
+    DPOStatisticsPageView, DPOStatisticsCSVView,
     UserStatisticsPageView, EmployeeView)
 from .views.report_views import (
     UserReportView, UserArchiveView,
@@ -23,37 +24,88 @@ from .views.scannerjob_views import ScannerjobListView, ScannerjobDeleteView
 from .views.manual_views import ManualMainView
 from .views.support_views import SupportButtonView
 
+reports_patterns = [
+    # Pages related to unhandled reports:
+    path("personal/",         UserReportView.as_view(),      name="personal"),
+    path("remediator/",       RemediatorView.as_view(),      name="remediator"),
+    path("undistributed/",    UndistributedView.as_view(),   name="undistributed"),
+    path("sbsys-personal/",   SBSYSPersonalView.as_view(),   name="sbsys-personal"),
+    path("sbsys-remediator/", SBSYSRemediatorView.as_view(), name="sbsys-remediator"),
+]
+
+archive_patterns = [
+    # Pages related to archived reports:
+    path("personal/",         UserArchiveView.as_view(),            name="personal"),
+    path("remediator/",       RemediatorArchiveView.as_view(),      name="remediator"),
+    path("undistributed/",    UndistributedArchiveView.as_view(),   name="undistributed"),
+    path("sbsys-personal/",   SBSYSPersonalArchiveView.as_view(),   name="sbsys-personal"),
+    path("sbsys-remediator/", SBSYSRemediatorArchiveView.as_view(), name="sbsys-remediator"),
+]
+
 urlpatterns = [
-    re_path(r'^$',      UserReportView.as_view(),     name="index"),
-    re_path(r'^reports$', UserReportView.as_view(), name="reports"),
-    re_path(r'^remediator$', RemediatorView.as_view(), name="remediator"),
-    re_path(r'^undistributed$', UndistributedView.as_view(), name="undistributed"),
-    re_path(r'^archive/reports', UserArchiveView.as_view(), name="reports-archive"),
-    re_path(r'^archive/remediator', RemediatorArchiveView.as_view(), name="remediator-archive"),
-    re_path(r'^archive/undistributed', UndistributedArchiveView.as_view(),
-         name="undistributed-archive"),  # noqa
-    re_path(r'^sbsys/personal', SBSYSPersonalView.as_view(), name="sbsys-personal"),
-    re_path(r'^sbsys/remediator', SBSYSRemediatorView.as_view(), name="sbsys-remediator"),
-    re_path(r'^sbsys/archive/personal', SBSYSPersonalArchiveView.as_view(),
-         name="sbsys-archive-personal"),   # noqa
-    re_path(r'^sbsys/archive/remediator', SBSYSRemediatorArchiveView.as_view(),
-         name="sbsys-archive-remediator"),   # noqa
-    re_path('api$',     JSONAPIView.as_view(),     name="json-api"),
-    path('account/<uuid:pk>', AccountView.as_view(), name="account"),
-    path('account/', AccountView.as_view(), name="account-me"),
-    path("account/outlook-category-settings/", AccountOutlookSettingView.as_view(),
-         name="outlook-category-settings"),
+    # Document Report views
+    path("",            UserReportView.as_view(),               name="index"),
+    path("reports/",    include((reports_patterns, "reports"),  namespace="reports")),
+    path("archive/",    include((archive_patterns, "archive"),  namespace="archive")),
+
+    # LEGACY --> NEW MAPPINGS:
+    # “/reports/” --> /reports/personal/
+    re_path(
+        r"^reports/?$",
+        RedirectView.as_view(
+            pattern_name="reports:personal",
+            permanent=True,
+            query_string=True
+        )
+    ),
+    # “/remediator/” --> /reports/remediator/
+    re_path(
+        r"^remediator/?$",
+        RedirectView.as_view(
+            pattern_name="reports:remediator",
+            permanent=True,
+            query_string=True
+        )
+    ),
+    # “/undistributed/” --> /reports/undistributed/
+    re_path(
+        r"^undistributed/?$",
+        RedirectView.as_view(
+            pattern_name="reports:undistributed",
+            permanent=True,
+            query_string=True
+        )
+    ),
+    # "/archive/reports/" --> /archive/personal/
+    re_path(
+        r"^archive/reports/?$",
+        RedirectView.as_view(
+            pattern_name="archive:personal",
+            permanent=True,
+            query_string=True
+        )
+    ),
+
+    # Scannerjob view
+    path('scannerjobs/', ScannerjobListView.as_view(), name="scannerjobs"),
+    path('scannerjobs/<int:pk>/delete', ScannerjobDeleteView.as_view(), name="delete_scannerjob"),
+
+    # Statistics views
     re_path(r'^statistics/leader/$', LeaderStatisticsPageView.as_view(), name='statistics-leader'),
     re_path(r'^statistics/leader/csv/$', LeaderStatisticsCSVView.as_view(), name='statistics-leader-export'),  # noqa
     re_path(r'^statistics/dpo/$', DPOStatisticsPageView.as_view(), name='statistics-dpo'),
     re_path(r'^statistics/dpo/csv/$', DPOStatisticsCSVView.as_view(), name='statistics-dpo-export'),
-    path('statistics/user/', UserStatisticsPageView.as_view(),
-         name='statistics-user-me'),
-    path('statistics/user/<uuid:pk>', UserStatisticsPageView.as_view(),
-         name='statistics-user-id'),
+    path('statistics/user/', UserStatisticsPageView.as_view(), name='statistics-user-me'),
+    path('statistics/user/<uuid:pk>', UserStatisticsPageView.as_view(), name='statistics-user-id'),
     path('statistics/employee/<uuid:pk>', EmployeeView.as_view(), name='employee'),
-    path('scannerjobs/', ScannerjobListView.as_view(), name="scannerjobs"),
-    path('scannerjobs/<int:pk>/delete', ScannerjobDeleteView.as_view(), name="delete_scannerjob"),
+
+    # Account view
+    path('account/<uuid:pk>', AccountView.as_view(), name="account"),
+    path('account/', AccountView.as_view(), name="account-me"),
+    path("account/outlook-category-settings/", AccountOutlookSettingView.as_view(),
+         name="outlook-category-settings"),
+
+    re_path('api$',     JSONAPIView.as_view(),     name="json-api"),
     re_path(r'^health/', lambda r: HttpResponse()),
     re_path(r'^version/?$', lambda r: HttpResponse(__version__)),
     re_path(r'^help/$', ManualMainView.as_view(), name="guide"),